Moved DimensionsEditor to Components
